parquet-pyarrow-integration-testing/requirements.txt
maturin
toml
pytest
pyarrow
pycryptodome
